requirements.txt
Django==1.7.4
Pygments==2.0.2
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.0.5
gunicorn==19.2.1
psycopg2==2.6
requests==2.5.1
six==1.9.0
static3==0.5.1
whitenoise==1.0.6
wsgiref==0.1.2
jsonfield==1.0.3
django-reversion==1.8.5
Markdown==2.6
django-extensions==1.5.0
python-dateutil==2.4.1
django-taggit==0.12.3
-e git+https://github.com/kobotoolbox/pyxform.git@f2bfc6c968ffec4f5128b9996ccefca3b5ac80d1#egg=pyxform-origin/koboform-stable
django-mptt==0.6.1
django-debug-toolbar==1.2.2
shortuuid==0.4.2

